unified vision.encoders with core.wrappers api
