+++ configs/vision/dino_vit/offline/monusac.yaml
@@ -32,7 +32,7 @@ trainer:
             0: train
             1: test
           backbone:
-            class_path: eva.vision.models.networks.encoders.TimmEncoder
+            class_path: eva.vision.models.wrappers.TimmModel
             init_args:
               model_name: ${oc.env:TIMM_MODEL_NAME, vit_small_patch16_224_dino}
               pretrained: ${oc.env:MODEL_PRETRAINED, true}
