+++ src/eva/core/models/__init__.py
@@ -1,7 +1,8 @@
 """Models API."""
 
 from eva.core.models.modules import HeadModule, InferenceModule
-from eva.core.models.networks import MLP, HuggingFaceModel, ModelFromFunction, ONNXModel
+from eva.core.models.networks import MLP
+from eva.core.models.wrappers import HuggingFaceModel, ModelFromFunction, ONNXModel
 
 __all__ = [
     "HeadModule",
