+++ src/eva/core/models/networks/__init__.py
@@ -1,6 +1,5 @@
 """Networks API."""
 
 from eva.core.models.networks.mlp import MLP
-from eva.core.models.networks.wrappers import HuggingFaceModel, ModelFromFunction, ONNXModel
 
-__all__ = ["MLP", "ModelFromFunction", "HuggingFaceModel", "ONNXModel"]
+__all__ = ["MLP"]
