--- src/eva/core/models/networks/transforms/__init__.py
@@ -1,6 +1,0 @@
-"""Model outputs transforms API."""
-
-from eva.core.models.networks.transforms.extract_cls_features import ExtractCLSFeatures
-from eva.core.models.networks.transforms.extract_patch_features import ExtractPatchFeatures
-
-__all__ = ["ExtractCLSFeatures", "ExtractPatchFeatures"]
