+++ src/eva/core/models/transforms/__init__.py
@@ -1,0 +1,6 @@
+"""Model outputs transforms API."""
+
+from eva.core.models.transforms.extract_cls_features import ExtractCLSFeatures
+from eva.core.models.transforms.extract_patch_features import ExtractPatchFeatures
+
+__all__ = ["ExtractCLSFeatures", "ExtractPatchFeatures"]
