+++ src/eva/core/models/wrappers/__init__.py
@@ -1,0 +1,13 @@
+"""Model Wrappers API."""
+
+from eva.core.models.wrappers.base import BaseModel
+from eva.core.models.wrappers.from_function import ModelFromFunction
+from eva.core.models.wrappers.huggingface import HuggingFaceModel
+from eva.core.models.wrappers.onnx import ONNXModel
+
+__all__ = [
+    "BaseModel",
+    "ModelFromFunction",
+    "HuggingFaceModel",
+    "ONNXModel",
+]
