+++ src/eva/vision/models/__init__.py
@@ -1,5 +1,5 @@
 """Vision Models API."""
 
-from eva.vision.models import networks
+from eva.vision.models import networks, wrappers
 
-__all__ = ["networks"]
+__all__ = ["networks", "wrappers"]
