+++ src/eva/vision/models/networks/__init__.py
@@ -1,7 +1,6 @@
 """Vision Networks API."""
 
-from eva.vision.models.networks import postprocesses
 from eva.vision.models.networks.abmil import ABMIL
-from eva.vision.models.networks.from_timm import TimmModel
+from eva.vision.models.wrappers.from_timm import TimmModel
 
-__all__ = ["postprocesses", "ABMIL", "TimmModel"]
+__all__ = ["ABMIL", "TimmModel"]
