--- src/eva/vision/models/networks/postprocesses/__init__.py
@@ -1,5 +1,0 @@
-"""Model post-process transforms."""
-
-from eva.vision.models.networks.postprocesses.cls import ExtractCLSFeatures
-
-__all__ = ["ExtractCLSFeatures"]
