+++ src/eva/vision/models/wrappers/__init__.py
@@ -1,0 +1,5 @@
+"""Vision Model Wrappers API."""
+
+from eva.vision.models.wrappers.from_timm import TimmModel
+
+__all__ = ["TimmModel"]
