--- tests/eva/vision/models/networks/encoders/__init__.py
@@ -1,1 +1,0 @@
-"""Test for vision encoders."""
