+++ tests/eva/vision/models/wrappers/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for vision model wrappers."""
